TGL-32.fix: Fixed types conflict
