+++ src/types/typeProfile.d.ts
@@ -1,0 +1,9 @@
+export interface ProfileMutation {
+  email: string;
+  firstName: string;
+  lastName: string;
+  middleName: string;
+  region: string;
+  settlement: string;
+  address: string;
+}
